+++ lib/utils/constant.dart
@@ -1,2 +1,5 @@
 const defaultCycleLength = 28; // Default Cycle length of menstrual cycle
 const defaultPeriodDuration = 5; // Default period duration of menstrual cycle
+const defaultFollicularDay =
+    6; // Default Follicular duration of menstrual cycle
+const defaultOvulationDay = 5; // Default ovulation duration of menstrual cycle