+++ lib/utils/strings.dart
@@ -1,20 +1,22 @@
 class Strings {
-  static String pkgUrl = "https://pub.dev/packages/menstrual_cycle_widget";
-  static String secretKeyLabel = "Must required secretKey for data encryption";
-  static String ivKeyLabel = "Must required ivKey for data encryption";
-  static String totalCycleDaysLabel =
+  static const String pkgUrl =
+      "https://pub.dev/packages/menstrual_cycle_widget";
+  static const String secretKeyLabel =
+      "Must required secretKey for data encryption";
+  static const String ivKeyLabel = "Must required ivKey for data encryption";
+  static const String totalCycleDaysLabel =
       "totalCycleDays size must be greater than 0";
-  static String totalPeriodDaysLabel =
+  static const String totalPeriodDaysLabel =
       "totalPeriodDays size must be greater than 0";
-  static String printLogLabel = "LOG FROM MENSTRUAL CYCLE";
-  static String saveLabel = "Save";
-  static String cancelLabel = "Cancel";
-  static String editPeriodDateLabel = "Edit Period dates";
-  static String periodLabel = "Period";
-  static String predictOvulationDateLabel = "Predict ovulation date";
-  static String predictPeriodDateLabel = "Predict period date";
-  static String selectedDateLabel = "Selected date";
-  static String todayLabel = "Today";
+  static const String printLogLabel = "LOG FROM MENSTRUAL CYCLE";
+  static const String saveLabel = "Save";
+  static const String cancelLabel = "Cancel";
+  static const String editPeriodDateLabel = "Edit Period dates";
+  static const String periodLabel = "Period";
+  static const String predictOvulationDateLabel = "Predict ovulation date";
+  static const String predictPeriodDateLabel = "Predict period date";
+  static const String selectedDateLabel = "Selected date";
+  static const String todayLabel = "Today";
   static const String menstruationLabel = "Menstruation";
   static const String follicularPhaseLabel = "Follicular phase";
   static const String ovulationLabel = "Ovulation";
@@ -22,17 +24,44 @@ class Strings {
   static const String dayLabel = "Day";
   static const String logPeriodLabel = "Log Period";
 
+  /// Messages for circle text
+  static const String pastCycleLabel = "Your period is";
+  static const String circleDaysLateLabel = "Days late";
+  static const String circleDayLateLabel = "Day late";
+  static const String lateLabel = "Log your period for better predictions";
+  static const String timeForPregnancyTestLabel = "Time for pregnancy test";
+  static const String timeForPregnancyTestTitleLabel = "Today &";
+  static const String timeForPregnancyTestTitleLabel2 = "more day";
+  static const String timeForPregnancyTestTitleLabel4 = "more days";
+  static const String timeForPregnancyTestTitleLabel3 = "Last day";
+  static const String predictPeriodLabel = "Predict period day: ";
+  static const String conceiveLabel = "Best chance to conceive is in";
+  static const String conceiveLabel2 = "Best chance to conceive is";
+  static const String periodPhaseLabel = "Period Day: ";
+  static const String phaseDayLabel = "Day";
+  static const String phaseDaysLabel = "Days";
+  static const String pregnancyChanceMsg1 = "Medium chance of getting pregnant";
+  static const String pregnancyChanceMsg2 = "High chance of getting pregnant";
+  static const String pregnancyChanceMsg3 = "Last fertile day of this cycle";
+  static const String pregnancyChanceMsg4 = "Low chance of getting pregnant";
+  static const String ovulationDayMsg1 = "Prediction";
+  static const String ovulationDayTitle = "Ovulation Day";
+  static const String predictPeriodDayStartLabel = "Predict period start is in";
+  static const String predictPeriodDayStartLabel2 =
+      "May be your period start will be";
+  static const String predictPeriodDayStartLabel1 = "Tomorrow";
+
   /// Error
-  static String errorInvalidSize =
+  static const String errorInvalidSize =
       "Invalid 'MenstrualCyclePhaseView' widget size. Size must be greater than or equal to 200. See ";
-  static String errorInvalidCount =
+  static const String errorInvalidCount =
       "Invalid 'totalCycleDays' count. Count must be greater than or equal to 10. See ";
-  static String errorInvalidMenstruationDayCount =
+  static const String errorInvalidMenstruationDayCount =
       "Invalid 'menstruationDayCount' count. Count must be greater than or equal to 0. See ";
-  static String errorInvalidFollicularDayCount =
+  static const String errorInvalidFollicularDayCount =
       "Invalid 'follicularDayCount' count. Count must be greater than or equal to 0. See ";
-  static String errorInvalidOvulationDayCount =
+  static const String errorInvalidOvulationDayCount =
       "Invalid 'ovulationDayCount' count. Count must be greater than or equal to 0. See ";
-  static String errorInvalidSymptomsLogDate =
+  static const String errorInvalidSymptomsLogDate =
       "Invalid symptoms log date. Date format is yyyy-MM-dd";
 }